Fix insruction size of old simulator

+++ compiler_set/simulator/logger.cpp
@@ -21,7 +21,7 @@ Logger::Logger(simulation_options * a_opt) {
 	}
 }
 
-void Logger::instruction(uint8_t pc, uint32_t inst) {
+void Logger::instruction(uint32_t pc, uint32_t inst) {
 	if (fp && opt->enable_record_instruction) {
 		fprintf(fp, "INST: %8d %08x\n", pc, inst);
 	}
